chore(@clayui/drop-down): update DropDown component stories
